rearrange where minimax specifics come in
