+++ npcpy/npc_sysenv.py
@@ -17,6 +17,7 @@ import requests
 ON_WINDOWS = platform.system() == "Windows"
 ON_MACOS = platform.system() == "Darwin"
 MINIMAX_MODELS = frozenset({"MiniMax-M3", "MiniMax-M2.7"})
+_PROVIDER_FALLBACK_MODELS = {"minimax": {"MiniMax-M2.7"}}
 
 def get_data_dir() -> str:
     """Get the data directory."""
@@ -148,7 +149,7 @@ def get_locally_available_models(project_directory, airplane_mode=False, gguf_di
 
         provider_model_attrs = {
             "anthropic": ("ANTHROPIC_API_KEY", "anthropic_models"),
-            "minimax": ("MINIMAX_API_KEY", None),
+            "minimax": ("MINIMAX_API_KEY", "minimax_models"),
             "moonshot": ("MOONSHOT_API_KEY", "moonshot_models"),
             "openai": ("OPENAI_API_KEY", "open_ai_chat_completion_models"),
             "gemini": ("GEMINI_API_KEY", "gemini_models"),
@@ -182,10 +183,8 @@ def get_locally_available_models(project_directory, airplane_mode=False, gguf_di
             if litellm is None:
                 continue
             try:
-                if attr_name is None:
-                    model_set = MINIMAX_MODELS
-                else:
-                    model_set = getattr(litellm, attr_name, None) or set()
+                litellm_models = getattr(litellm, attr_name, None) or set() if attr_name else set()
+                model_set = litellm_models | _PROVIDER_FALLBACK_MODELS.get(provider, set())
                 if not model_set:
                     continue
                 for model_id in model_set:
@@ -1012,10 +1011,6 @@ def lookup_provider(model: str) -> str:
             return "lora"
 
     if model in MINIMAX_MODELS:
-        return "minimax"
-
-    lowered = model.lower()
-    if lowered.startswith("minimax/") or lowered.startswith("minimax-"):
         return "minimax"
 
     if model == "deepseek-chat" or model == "deepseek-reasoner":
